Upgrade typescript-eslint to v8 and fix all lint errors
Bumps @typescript-eslint/eslint-plugin and @typescript-eslint/parser from
v5 to v8, fixes all lint errors/warnings to reach zero, and adds a lint
CI job so the repo can't silently drift again.

Supersedes Renovate PR #19.

Key changes:
- Extract enums from component files into .types.ts siblings to fix
  react-refresh/only-export-components warnings
- Convert method-style onChange signatures to property-style to fix
  @typescript-eslint/unbound-method
- Remove stale eslint-disable directives for non-existent rules
- Fix CookieSettingsLink unused-expressions with early return
- Narrow Link href.toString() to avoid no-base-to-string on UrlObject
- Add varsIgnorePattern/argsIgnorePattern for _-prefixed variables
- Defer Button/Input static-components and InputDatePicker hooks
  refactors with targeted eslint-disable + TODO comments
