+++ src/components/Alert/Alert.types.ts
@@ -1,0 +1,6 @@
+export enum AlertType {
+  Info = 'info',
+  Warning = 'warning',
+  Error = 'error',
+  Help = 'help',
+}
