+++ src/components/Button/Button.tsx
@@ -1,38 +1,13 @@
+/* eslint-disable react-hooks/static-components -- TODO: hoist ButtonContent out of Button render body. Deferred to a follow-up PR. */
 import React from 'react';
 import styles from './Button.module.scss';
 import classNames from 'classnames';
 import { Link } from '..';
 import { Brand } from 'components/Link/Link';
 import { Spinner } from 'components/Spinner/Spinner';
+import { ButtonVariant, ButtonSize } from './Button.types';
 
-export enum ButtonVariant {
-  Primary = 'primary',
-  PrimaryLight = 'primaryLight',
-  Secondary = 'secondary',
-  SecondaryBold = 'secondaryBold',
-  Transparent = 'transparent',
-  Outline = 'outline',
-  OutlineBold = 'outlineBold',
-  Sales = 'sales',
-  Naked = 'naked', // button without button styles or paddings
-  Success = 'success',
-  Alert = 'alert',
-  Bookmark = 'bookmark', // icon only button with special hover styles
-
-  // Used only with old (before 2025) Design System
-  Tertiary = 'tertiary',
-  Blend = 'blend',
-  Delete = 'delete',
-  Reject = 'reject',
-}
-
-export enum ButtonSize {
-  XS = 'extraSmall',
-  S = 'small',
-  M = 'medium',
-  // Used only with old (before 2025) Design System
-  L = 'large',
-}
+export { ButtonVariant, ButtonSize };
 
 export interface ButtonProps extends React.DetailedHTMLProps<
   React.ButtonHTMLAttributes<HTMLButtonElement>,
@@ -55,7 +30,7 @@ export interface ButtonProps extends React.DetailedHTMLProps<
   // Allow all props.
   // In remix, it's common for buttons to have names, values, etc.
   // Not to forget formAction. Typechecking for all of them doesn't serve much purpose since they're just strings.
-  [buttonExtraProp: string]: any;
+  [buttonExtraProp: string]: unknown;
 }
 
 export const Button = ({
@@ -96,6 +71,8 @@ export const Button = ({
     [className]: !moduleExtend,
   });
 
+  // TODO: hoist ButtonContent out of Button render body to satisfy
+  // react-hooks/static-components. Deferred to a follow-up PR.
   const ButtonContent = () => (
     <>
       {iconPosition === 'left' && icon}
