+++ src/components/Button/Button.types.ts
@@ -1,0 +1,28 @@
+export enum ButtonVariant {
+  Primary = 'primary',
+  PrimaryLight = 'primaryLight',
+  Secondary = 'secondary',
+  SecondaryBold = 'secondaryBold',
+  Transparent = 'transparent',
+  Outline = 'outline',
+  OutlineBold = 'outlineBold',
+  Sales = 'sales',
+  Naked = 'naked', // button without button styles or paddings
+  Success = 'success',
+  Alert = 'alert',
+  Bookmark = 'bookmark', // icon only button with special hover styles
+
+  // Used only with old (before 2025) Design System
+  Tertiary = 'tertiary',
+  Blend = 'blend',
+  Delete = 'delete',
+  Reject = 'reject',
+}
+
+export enum ButtonSize {
+  XS = 'extraSmall',
+  S = 'small',
+  M = 'medium',
+  // Used only with old (before 2025) Design System
+  L = 'large',
+}
