+++ src/components/Heading/Heading.tsx
@@ -1,26 +1,9 @@
 import styles from './Heading.module.scss';
 import classNames from 'classnames';
+import { HeadingStyle, HeadingLevel } from './Heading.types';
 
-export enum HeadingStyle {
-  ExtraLarge = 'extraLarge',
-  Large = 'large',
-  Default = 'default',
-  Small = 'small',
-  Tiny = 'tiny',
-  Secondary = 'secondary',
-  Subheading = 'subheading',
-  TeaserXL = 'teaserXl',
-  TeaserL = 'teaserL',
-  TeaserM = 'teaserM',
-  TeaserS = 'teaserS',
-  TeaserXS = 'teaserXs',
-  TeaserTopic = 'teaserTopic',
-  ArticleHeading = 'articleHeading',
-  ArticleSubheading = 'articleSubheading',
-  PlayerHeading = 'playerHeading',
-}
-
-export type HeadingLevel = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
+export { HeadingStyle };
+export type { HeadingLevel };
 
 interface HeadingProps {
   text: string | React.ReactNode;
