+++ src/components/Heading/Heading.types.ts
@@ -1,0 +1,20 @@
+export enum HeadingStyle {
+  ExtraLarge = 'extraLarge',
+  Large = 'large',
+  Default = 'default',
+  Small = 'small',
+  Tiny = 'tiny',
+  Secondary = 'secondary',
+  Subheading = 'subheading',
+  TeaserXL = 'teaserXl',
+  TeaserL = 'teaserL',
+  TeaserM = 'teaserM',
+  TeaserS = 'teaserS',
+  TeaserXS = 'teaserXs',
+  TeaserTopic = 'teaserTopic',
+  ArticleHeading = 'articleHeading',
+  ArticleSubheading = 'articleSubheading',
+  PlayerHeading = 'playerHeading',
+}
+
+export type HeadingLevel = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
