+++ src/components/Link/brand.ts
@@ -1,0 +1,5 @@
+export enum Brand {
+  MT = 'mt',
+  KV = 'kv',
+  VM = 'vm',
+}
