+++ src/components/ListHeading/ListHeading.types.ts
@@ -1,0 +1,7 @@
+export enum ListHeadingType {
+  Default = 'default',
+  NewsLarge = 'newsLarge',
+  NewsSmall = 'newsSmall',
+  AdLarge = 'adLarge',
+  AdSmall = 'adSmall',
+}
