+++ src/components/ListItem/ListItem.types.ts
@@ -1,0 +1,5 @@
+export enum ListItemType {
+  Default = 'default',
+  HorseAdItem = 'horseAdItem',
+  JobAdItem = 'jobAdItem',
+}
