+++ src/components/Tag/Tag.types.ts
@@ -1,0 +1,12 @@
+export enum TagType {
+  Plain = 'plain',
+  Buying = 'buying',
+  Selling = 'selling',
+  Leasing = 'leasing',
+  Premium = 'premium',
+  PremiumPlain = 'premiumPlain',
+  PremiumLight = 'premiumLight',
+  PremiumLightPlain = 'premiumLightPlain',
+  Primary = 'primary',
+  Outline = 'outline',
+}
